Fix WTF::copyLCharsFromUCharSource() to compile with -Wshorten-64-to-32 <http://webkit.org/b/107227>
Reviewed by Benjamin Poulain.

Fixes the following build error:

    ASCIIFastPath.h:117:59: error: implicit conversion loses integer precision: 'unsigned long' to 'const unsigned int' [-Werror,-Wshorten-64-to-32]
            const unsigned endLength = length - ucharsPerLoop + 1;
                           ~~~~~~~~~   ~~~~~~~~~~~~~~~~~~~~~~~^~~
    1 error generated.

* wtf/text/ASCIIFastPath.h:
(WTF::copyLCharsFromUCharSource): Change local variables from
unsigned to size_t.

Canonical link: https://commits.webkit.org/125524@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@140162 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WTF/wtf/text/ASCIIFastPath.h
@@ -112,9 +112,9 @@ inline void copyLCharsFromUCharSource(LChar* destination, const UChar* source, s
     }
 
     const uintptr_t sourceLoadSize = 32; // Process 32 bytes (16 UChars) each iteration
-    const unsigned ucharsPerLoop = sourceLoadSize / sizeof(UChar);
+    const size_t ucharsPerLoop = sourceLoadSize / sizeof(UChar);
     if (length > ucharsPerLoop) {
-        const unsigned endLength = length - ucharsPerLoop + 1;
+        const size_t endLength = length - ucharsPerLoop + 1;
         for (; i < endLength; i += ucharsPerLoop) {
 #ifndef NDEBUG
             for (unsigned checkIndex = 0; checkIndex < ucharsPerLoop; ++checkIndex)
